requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
browsermob-proxy==0.8.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.13.0
fonttools==4.43.1
fsspec==2023.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.6.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
radian==0.6.7
rchitect==0.4.2
requests==2.31.0
selenium==4.13.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
torch==2.1.0
torchvision==0.16.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==3.0.0
wsproto==1.2.0
XlsxWriter==3.1.5
zstandard==0.21.0
